requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pandas>=2.0.0
llama-index>=0.10.0
llama-index-embeddings-openai>=0.1.0
llama-index-llms-openai>=0.1.0
llama-index-llms-anthropic>=0.1.0
llama-index-vector-stores-chroma>=0.1.0
chromadb>=0.4.0
openai>=1.0.0
anthropic>=0.18.0
ddgs>=9.0.0
